Add menu for pause screen
When the player pauses during gameplay, they will be presented with a three
button menu. The buttons are "resume", "end game", and "quit". "Resume" is self-
explanatory. "End game" ends the game and returns to the menu screen. "Quit" ends
the game and closes the game engine, returning the user to their desktop.